NXP-28146: Remove unused code
